Add `call_with_list` to `IdaCaller`
It's awkward to use the CSV interface to pass around normal lists, i.e.
ones where each item is a primitive, e.g. a number, rather than a dict.
With CSVs, we need to wrap each item in a row, etc. Now, it just writes
and reads each value on its own line.